Abstract out common logic to generate patched test suites.
