+++ test/test_dynamic_shapes.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env pytest
-import functools
-from unittest.mock import patch
 
-import torchdynamo
+from torchdynamo.testing import make_test_cls_with_patches
 
 from . import test_functions
 from . import test_misc
@@ -11,33 +9,8 @@ from . import test_repros
 from . import test_unspec
 
 
-def make_dynamic_fn(fn):
-    @functools.wraps(fn)
-    def _fn(*args, **kwargs):
-        with patch.object(torchdynamo.config, "dynamic_shapes", True):
-            return fn(*args, **kwargs)
-
-    return _fn
-
-
 def make_dynamic_cls(cls):
-    class DynamicShapeTest(cls):
-        pass
-
-    DynamicShapeTest.__name__ = f"DynamicShapes{cls.__name__}"
-
-    for name in dir(cls):
-        if name.startswith("test_"):
-            fn = getattr(cls, name)
-            if not callable(fn):
-                continue
-            new_name = f"{name}_dynamic_shapes"
-            fn = make_dynamic_fn(fn)
-            fn.__name__ = new_name
-            setattr(DynamicShapeTest, name, None)
-            setattr(DynamicShapeTest, new_name, fn)
-
-    return DynamicShapeTest
+    return make_test_cls_with_patches(cls, "DynamicShapes", "_dynamic_shapes", ("dynamic_shapes", True))
 
 
 DynamicShapesFunctionTests = make_dynamic_cls(test_functions.FunctionTests)
